Add files via upload
Libraries and Modules: It begins by importing necessary libraries such as NLTK, which is used for text manipulation and preprocessing in Python. Specific modules for removing stopwords and calculating cosine distance are loaded, indicating tasks related to text processing and similarity calculations.

Text Processing: The notebook handles text data by opening and reading a file named "Text1.txt", which contains one or more paragraphs of text. The text is split into sentences, which are further processed to remove non-alphabetic characters and split into words. This indicates that the notebook might be preparing the data for further NLP tasks.

Function for Sentence Similarity: A function is defined to calculate the similarity between two sentences based on cosine distance of their vector representations. This involves converting sentences into vectors where each position corresponds to a word in the combined vocabulary of both sentences, and the value is the frequency of that word in the sentence.

Similarity Matrix Calculation: The notebook calculates a similarity matrix for the sentences extracted from the text file. This matrix helps in understanding how each sentence is related to every other sentence in terms of content, which can be crucial for tasks such as text summarization or clustering.